heatDiff_int_2D: kept working on matrix directory

+++ heatDiff_int_2D/matrix/matrixStruct.h
@@ -8,9 +8,9 @@ struct Matrix {
 
     std::vector<double> val;
 
-    std::vector<double> col;
+    std::vector<int> col;
 
-    std::vector<double> poi;
+    std::vector<int> poi;
 
     std::vector<double> A;
